_generator/requirements.txt
python-liquid[autoescape]
